requirements.txt
alpaca-trade-api>=3.0.0
pandas>=2.2.3
numpy>=2.2.6
pyyaml>=6.0.1
fastapi>=0.115.1
uvicorn>=0.34.3
scikit-learn>=1.6.1
requests>=2.32.3
